hustle-backend/requirements.txt
asgiref==3.5.0
backports.zoneinfo==0.2.1
certifi==2021.10.8
charset-normalizer==2.0.12
cloudinary==1.29.0
dj3-cloudinary-storage==0.0.6
Django==4.0.4
django-cors-headers==3.11.0
django-filter==21.1
djangorestframework==3.13.1
djangorestframework-simplejwt==5.1.0
idna==3.3
importlib-metadata==4.11.3
Markdown==3.3.6
Pillow==9.1.0
psycopg2==2.9.3
psycopg2-binary==2.9.3
PyJWT==2.3.0
python-decouple==3.6
pytz==2022.1
requests==2.27.1
six==1.16.0
sqlparse==0.4.2
urllib3==1.26.9
zipp==3.8.0
